// arcadia.backoffice.fe/src/modules/dashboard/types.ts
export enum MachinesStatusColor {
  error = '#e02020',
  offline = '#d8d8d8',
  shuttingDown = '#474749',
  drying = '#fa6400',
  preparing = '#32c5ff',
  inPlay = '#6DD400',
  seeding = '#F7B500',
  ready = '#0091ff',
}

export enum MachinesStatusLabel {
  error = 'Error',
  offline = 'Offline',
  shuttingDown = 'Shutting Down',
  drying = 'Drying',
  preparing = 'Preparing',
  inPlay = 'In Play',
  seeding = 'Seeding',
  ready = 'Ready',
}

export interface IMachinesStatusChartItem {
  name: MachinesStatusLabel
  color: MachinesStatusColor
  value: number
}

export type IMachinesStatusChartData = IMachinesStatusChartItem[];